+++ changelog.md
@@ -1,5 +1,9 @@
 # Payplug module changelog
 
+## Next Version
+- Feature :
+  - [PRE-2159](https://git.payplug.com/plugins/prestashop_v2_1.7/-/merge_requests/2370): Add applepay checkout button on cart page
+
 ## Version 4.7.0
 - Refactoring :
   - [PRE-1985](https://git.payplug.com/plugins/prestashop_v2_1.7/-/merge_requests/2309): Add action Order